Do more thready stuff Process manager modified to prevent deadlock
