--- src/main/java/com/cascadebot/cascadewrapper/ProcessThread.java
@@ -1,10 +1,0 @@
-package com.cascadebot.cascadewrapper;
-
-public class ProcessThread extends Thread {
-
-    @Override
-    public void run() {
-
-    }
-
-}
