+++ src/main/java/com/cascadebot/cascadewrapper/Util.java
@@ -1,5 +1,7 @@
 package com.cascadebot.cascadewrapper;
 
+import com.cascadebot.shared.SharedConstants;
+
 public class Util {
 
     public static <T extends Enum<T>> T getSafeEnum(Class<T> enumClass, String value) {
@@ -8,6 +10,10 @@ public class Util {
         } catch (IllegalArgumentException e) {
             return null;
         }
+    }
+
+    public static String getBotCommand(String command, String[] args) {
+        return SharedConstants.BOT_OP_PREFIX + " " + command + String.join(" ", args);
     }
 
 }
